Research paper published
This paper presents a lithium-ion battery charger based on microcontroller that primarily works on constant current, constant voltage (CC/CV) type of charging algorithm. Lithium-ion batteries outperform conventional battery technology in terms of longevity and power density, providing longer battery life in a lighter container. But in order to make the best use out of this battery, there is a huge requirement of such a charger that is specifically designed to observe the battery’s charging status and provide a charge, as soon as necessary, in order to maintain the battery, keeping it healthy and energized. The following prototype can be applicable in electric vehicles etc. It completely eliminates the monitoring part for the users towards the battery and allow them to multitask or else focus on different activities.